Interactive Selection Widget

Construct control widget

interactive.selection(action, nodes, edges = NULL, slider = NULL,
  checkbox = NULL, perturb = T, monitor = T)

Arguments

action
function to perform the widgets action

nodes
node labels

edges
edge labels

slider
slider label

checkbox
checkbox label

perturb
should a node perturbation control be rendered

monitor
should a node monitoring control be rendered

Details

Constructs a toplevel window that allows the allowing the user to interactively select nodes to perturb/monitor, from a subset of models, and then perform a given action.

The action argument must be function of five arguments

perturb the nodes that were perturbed

monitor the outcome of the monitoring

edge the edges to select

check the state of a checkbutton

slider the state of a slider